requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt
python-decouple
dj-database-url
requests